api/pyproject.toml
[project]
name = "deployment-lab"
version = "0.1.0"
description = ""
authors = [
    {name = "Abdullah Othman"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "docker>=6.0.0",
    "pydantic>=2.0.0",
    "motor>=3.0.0",
    "pymongo>=4.0.0",
    "aiofiles",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "cloudflare",
    "gitpython",
    "jinja2",
    "python-dotenv"
]

[tool.poetry]
package-mode=false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
